refactor: extract remaining e2e_support StatefulWidget pin hosts (#4598)
Move _DismissibleSeaDialog and _DisappearingLabel into test/support so pin
suites no longer declare leftover widget hosts, and drop the stray duplicate
AC1 alias doc on enterStandardE2eScenario.
