+++ packages/colonizethis_app_e2e_support/test/e2e_idle_poll_step_test.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:colonizethis_app_e2e_support/e2e_test_shared.dart';
+import 'support/disappearing_label_host.dart';
 import 'support/e2e_widget_pump_harness.dart';
 
 void main() {
@@ -31,7 +32,7 @@ void main() {
     testWidgets('pumps until finder clears then returns', (
       WidgetTester tester,
     ) async {
-      await tester.pumpWidget(wrapE2eApp(_DisappearingLabel()));
+      await tester.pumpWidget(wrapE2eApp(const DisappearingLabelHost()));
       expect(find.text('gone'), findsOneWidget);
       await tester.tap(find.text('hide'));
       await e2ePumpUntilFinderEmpty(
@@ -51,30 +52,4 @@ void main() {
       expect(e2eNextIdlePollStepMs(500), 500);
     });
   });
-}
-
-class _DisappearingLabel extends StatefulWidget {
-  const _DisappearingLabel();
-
-  @override
-  State<_DisappearingLabel> createState() => _DisappearingLabelState();
-}
-
-class _DisappearingLabelState extends State<_DisappearingLabel> {
-  bool _showGone = true;
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Column(
-        children: [
-          if (_showGone) const Text('gone'),
-          TextButton(
-            onPressed: () => setState(() => _showGone = false),
-            child: const Text('hide'),
-          ),
-        ],
-      ),
-    );
-  }
 }
